contributors/ttwf/jackyy/incoming/css-transforms/transform_translate_invalid_prefixed.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate with invalid translation value</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css3-transforms/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css3-transforms/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="Check if transform sets translate(null, null) that an expection is to be thrown"> 
    <script type="text/javascript" src="../../resources/testharness.js"></script>
    <script type="text/javascript" src="../../resources/testharnessreport.js"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
        // Set the transform 
        document.getElementById("test").style.webkitTransform = "translate(null, null)";
        /* document.getElementById("test").style.mozkitTransform = "translate(null, null)";
        document.getElementById("test").style.msTransform = "translate(null, null)";
        document.getElementById("test").style.oTransform = "translate(null, null)";
        document.getElementById("test").style.transform = "translate(null, null)"; */
 
        // Verify that the transform was set as expected
        var result;
        var value;
        try {
            value = document.getElementById("test").style.getPropertyCSSValue("-webkit-transform").cssText;
            /* value = document.getElementById("test").style.getPropertyCSSValue("-moz-transform").cssText;
            value = document.getElementById("test").style.getPropertyCSSValue("-ms-transform").cssText;
            value = document.getElementById("test").style.getPropertyCSSValue("-0-transform").cssText;
            value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText; */
            result = false;
        } catch (e) {
            result = true;
            value = e.message;
        }
        
        test(function() {assert_true(result,
                        "transform should be translate(null, null)")}, 
                        "transform_translate_null_null");
    </script>    
</body>
</html>
